services/tasks_api/pyproject.toml
[tool.poetry]
name = "tasks-api"
version = "0.1.0"
description = ""
authors = ["Kentaro Nakamaye <kentaro.nakamae@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
httpx = "^0.25.0"
mangum = "^0.17.0"
boto3 = "^1.28.62"
pyjwt = "^2.8.0"
fastapi = "^0.99"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
bandit = "^1.7.5"
moto = "^4.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
